Create getOrCreateUser service for authentication
A new service function named getOrCreateUser has been added for user authentication. This function, used within the authorization process, fetches the user from Linear, and encrypts the user's access token, after which the user is upserted in the database. In the authorization service, instead of directly using the access token, getUserOrCreate is invoked to obtain the user's encrypted token.