ACEB-5: Configure Prisma (#5)
* Add Prisma

Add prisma dependency and configuration

* Rename configuration error file

* Rename configuration error file

* Consolidate errors

* Remove files from template init

* Reformat server for clarity and to handle removed Paths concept

* Remove logger and hello world route

* Create getOrCreateUser service for authentication

A new service function named getOrCreateUser has been added for user authentication. This function, used within the authorization process, fetches the user from Linear, and encrypts the user's access token, after which the user is upserted in the database. In the authorization service, instead of directly using the access token, getUserOrCreate is invoked to obtain the user's encrypted token.

* Allow console logging

* Frontend relies on `accessToken`

* Add tests, improve error logging

* Mock User

* Mock User

* Generate prisma client

* Correct schema patth

* Remove extraneous newline